requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-ai-generativelanguage